Refactor: fix all css mistakes

+++ src/renderer/components/Messages/Input/Field/Autosize.ts
@@ -4,11 +4,12 @@ export default {
 
     inserted: function (el: Element) {
         const style = getComputedStyle(el);
-        const minHeight = parseFloat(style.getPropertyValue("min-height"));
-        const maxHeight = parseFloat(style.getPropertyValue("max-height"));
+        const minHeight = parseInt(style.getPropertyValue("min-height"));
+        const maxHeight = parseInt(style.getPropertyValue("max-height"));
   
         const resize = () => {
             el.style.height = "auto";
+
             const height = Math.max(Math.min(el.scrollHeight, maxHeight), minHeight);
             el.style.setProperty("overflow-y", height < maxHeight ? "hidden" : "overlay");
             el.style.setProperty("height", `${height}px`, "important");
